pyproject.toml
[tool.poetry]
name = "debt-calculator"
version = "0.1.0"
description = ""
authors = ["AbdelrhmanHamouda <abdelrhman.ahmed@outlook.com>"]

[tool.poetry.dependencies]
python = "~3.9"
pydantic = "^1.8.2"
python-dateutil = "^2.8.2"
matplotlib = "^3.5.0"
scipy = "^1.7.3"

[tool.poetry.dev-dependencies]
coverage = "^6.2"
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
